Modified assignment2.c and Added Makefile
I added a make file so that the file compiles to the name monte.
I deleted the old assignment2.c file and wrote the new one from the beginning just to make sure that its clean, please go through the code and see if it has any issues. It runs fine on my machine.

Signed-off-by: Asif Rasheed <asif@linuxmail.org>